requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2 